pyproject.toml
[tool.poetry]
name = "codematic"
version = "0.1.0"
description = ""
authors = ["Pamilerin Idowu <pamilerinid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
